pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-based-chatbot"
version = "0.1.0"
description = "A RAG-based chatbot service"
requires-python = ">=3.9"

dependencies = [
    "fastapi[standard]==0.121.2",
    "langchain==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-community==0.3.31",
    "langchain-openai==0.3.35",
    "python-dotenv==1.0.0",
    "tiktoken>=0.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]
